create EventRestController, move there some methods
